fixed minor errors (use delete[])

+++ utils.cc
@@ -11,8 +11,8 @@
 /* --------------------------------------------------------------------
  * Utility function implementations
  *
- * $Date: 2007-08-02 12:54:35 $
- * $Revision: 1.7 $
+ * $Date: 2007-08-07 16:58:29 $
+ * $Revision: 1.8 $
  * -------------------------------------------------------------------- */
 
 
@@ -278,15 +278,17 @@ CxxUtil::utoa (unsigned long num)
   snprintf(ctemp, sizeof ctemp, "%lu", num);
   return std::string(ctemp);
 }
+#include <iostream>
 
 double
 CxxUtil::atod(const std::string &str)
 {
   char *error = NULL;
-  long ret = (int)strtod(str.c_str(), &error);
+  double ret = strtod(str.c_str(), &error);
   if (*error != '\0')
     {
-      throw std::invalid_argument(std::string("Invalid number format at: ") + error);
+      throw std::invalid_argument(std::string("Invalid number format at: ")
+				  + error);
     }
   return ret;
 }
